refactor(budget-details): migrate UI to grid layout with improved structure
