+++ frontend/projects/webapp/src/app/feature/budget/budget-details/budget-grid/index.ts
@@ -1,0 +1,8 @@
+export { BudgetGrid } from './budget-grid';
+export { BudgetGridCard } from './budget-grid-card';
+export { BudgetGridMobileCard } from './budget-grid-mobile-card';
+export { BudgetGridSection } from './budget-grid-section';
+export {
+  BudgetDetailPanel,
+  type BudgetDetailPanelData,
+} from './budget-detail-panel';
