--- frontend/projects/webapp/src/app/feature/budget/budget-details/budget-table/budget-table-view-mode.ts
@@ -1,1 +1,0 @@
-export type BudgetTableViewMode = 'envelopes' | 'table';
