+++ frontend/projects/webapp/src/app/feature/budget/budget-details/budget-table/index.ts
@@ -1,0 +1,1 @@
+export { BudgetTableView } from './budget-table-view';
