+++ frontend/projects/webapp/src/app/feature/budget/budget-details/components/index.ts
@@ -1,0 +1,4 @@
+export { BudgetProgressBar } from './budget-progress-bar';
+export { BudgetKindIndicator } from './budget-kind-indicator';
+export { BudgetActionMenu } from './budget-action-menu';
+export { BudgetViewToggle } from './budget-view-toggle';
