+++ frontend/projects/webapp/src/app/feature/budget/budget-details/data-core/budget-item-constants.ts
