+++ frontend/projects/webapp/src/app/feature/budget/budget-details/data-core/budget-view-mode.ts
@@ -1,0 +1,1 @@
+export type BudgetViewMode = 'envelopes' | 'table';
