+++ frontend/projects/webapp/src/app/feature/budget/budget.routes.ts
@@ -1,7 +1,6 @@
 import { type Routes } from '@angular/router';
 import { PAGE_TITLES } from '@core/routing';
 import { BudgetLineApi } from './budget-details/budget-line-api/budget-line-api';
-import { BudgetTableDataProvider } from './budget-details/budget-table/budget-table-data-provider';
 import { BudgetListStore } from './budget-list/budget-list-store';
 
 export const budgetRoutes: Routes = [
@@ -18,7 +17,7 @@ export const budgetRoutes: Routes = [
         path: ':id',
         title: PAGE_TITLES.BUDGET_DETAILS,
         data: { breadcrumb: 'Détail du budget', icon: 'visibility' },
-        providers: [BudgetLineApi, BudgetTableDataProvider],
+        providers: [BudgetLineApi],
         loadComponent: () => import('./budget-details/budget-details-page'),
       },
     ],
